refactor: cleanup dead code, unused deps, and error logging
Remove dead EndReason::child_signal and its now-unused ExitStatusExt
import; drop unused tracing/tracing-subscriber deps (all diagnostics go
via eprintln!). Log prompt-writer and run-log-writer errors instead of
swallowing them. Narrow Verdict.duration_ms to u64 for consumer
compatibility. Make opencode::extract_result fall back to an earlier
valid text event if the trailing one is malformed. Add explanatory
comments on ignored-signal inheritance (trap-SIGTERM test) and the
reader-task panic-free invariant.

+++ crates/cast-agent/src/finalize.rs
@@ -76,7 +76,7 @@ pub struct Verdict {
     pub log_path: String,
     pub prompt_path: String,
     pub event_count: usize,
-    pub duration_ms: u128,
+    pub duration_ms: u64,
     /// A short human-readable failure reason when one is available (spawn /
     /// supervision failure). Omitted from `result.json` when absent.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -195,7 +195,7 @@ pub fn build_verdict(
         log_path: log_path.to_string_lossy().into_owned(),
         prompt_path: prompt_path.to_string_lossy().into_owned(),
         event_count: events.len(),
-        duration_ms: duration.as_millis(),
+        duration_ms: duration.as_millis() as u64,
         error_detail,
         interrupt_signal,
     }
